requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
awscli==1.22.50
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
boto3==1.20.50
botocore==1.23.50
bs4==0.0.1
cffi==1.15.0
click==8.0.3
colorama==0.4.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.15.2
entrypoints==0.4
executing==0.8.2
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rsa==4.7.2
s3transfer==0.5.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
